Build and Documentation Improvements
Add descriptions to specific variables in the book data structures. Updated the settings ui to include units on a seperate column.  Restructured the mainwindow ui to be more space efficent. Fixed a issue with the settings dialog not reseting the changed settings on "reset." Chnaged the formating of confirmation books and other small details.

+++ bookdialog.cpp
@@ -59,22 +59,25 @@ void BookDialog::onDoubleClicked(const QModelIndex &index)
 void BookDialog::onDelete()
 {
     auto selected = ui->treeView->selectionModel()->selection().indexes();
-    QString str("Are you sure you want to delete those book(s)?\n");
-    QString bookstr{};
-    std::vector<int> bks{};
-    for(QModelIndex index : selected)
+    if(selected.size() > 0)
     {
-        int id = sortModel->data(index, Qt::DisplayRole).toInt();
-        QString bk = QString("Book %1 \n").arg(id);
-        bks.push_back(id);
-        bookstr.append(bk);
-    }
-    str.append(bookstr);
+        QString str("Are you sure you want to delete those book(s)?\n");
+        QString bookstr{};
+        std::vector<int> bks{};
+        for(QModelIndex index : selected)
+        {
+            int id = sortModel->data(index, Qt::DisplayRole).toInt();
+            QString bk = QString("Book %1\n").arg(id);
+            bks.push_back(id);
+            bookstr.append(bk);
+        }
+        str.append(bookstr);
 
-    ConfirmDialog dialog(str);
-    if(dialog.exec())
-    {
-        emit deleteBooks(bks);
-        accept();
+        ConfirmDialog dialog(str, "Delete Book?");
+        if(dialog.exec())
+        {
+            emit deleteBooks(bks);
+            accept();
+        }
     }
 }
